+++ src/components/PupilJourneyComponents/QuizUtils/answerTypeDiscriminators.ts
@@ -1,13 +1,10 @@
+// these discriminators should probably be moved outside of the components folder so that they can be consumed across the app
+
 import {
   AnswersSchema,
   MCAnswer,
   ShortAnswer,
 } from "@/node-lib/curriculum-api-2023/shared.schema";
-import { QuizMCQMultiAnswer } from "@/components/PupilJourneyComponents/QuizMCQMultiAnswer";
-import { QuizMCQSingleAnswer } from "@/components/PupilJourneyComponents/QuizMCQSingleAnswer";
-import { QuizShortAnswer } from "@/components/PupilJourneyComponents/QuizShortAnswer";
-
-// these discriminators should probably be moved outside of the components folder so that they can be consumed across the app
 
 export const isMultiAnswerMCQ = (
   answers: AnswersSchema,
@@ -31,17 +28,4 @@ export const isShortAnswer = (
   answers: AnswersSchema,
 ): answers is { "short-answer": ShortAnswer[] } => {
   return !!answers["short-answer"];
-};
-
-export const pickAnswerComponent = (answers: AnswersSchema) => {
-  switch (true) {
-    case isMultiAnswerMCQ(answers):
-      return QuizMCQMultiAnswer;
-    case isSingleAnswerMCQ(answers):
-      return QuizMCQSingleAnswer;
-    case isShortAnswer(answers):
-      return QuizShortAnswer;
-    default:
-      return null;
-  }
 };
