+++ src/components/PupilJourneyComponents/QuizUtils/pickAnswerComponent.test.ts
@@ -1,8 +1,4 @@
-import {
-  isMultiAnswerMCQ,
-  isShortAnswer,
-  pickAnswerComponent,
-} from "./pickAnswerComponent";
+import { pickAnswerComponent } from "./pickAnswerComponent";
 
 import { QuizMCQMultiAnswer } from "@/components/PupilJourneyComponents/QuizMCQMultiAnswer";
 import { QuizMCQSingleAnswer } from "@/components/PupilJourneyComponents/QuizMCQSingleAnswer";
@@ -36,36 +32,6 @@ describe("pickAnswerComponent", () => {
     "short-answer": [{}],
   } as AnswersSchema;
 
-  describe("isMultiAnswerMCQ", () => {
-    it("should return true if there are more than one correct answers", () => {
-      const result = isMultiAnswerMCQ(multiAnswer);
-      expect(result).toBe(true);
-    });
-    it("should return false if there is one correct answer", () => {
-      const result = isMultiAnswerMCQ(singleAnswer);
-      expect(result).toBe(false);
-    });
-  });
-  describe("isSingleAnswerMCQ", () => {
-    it("should return true if there is one correct answer", () => {
-      const result = isMultiAnswerMCQ(singleAnswer);
-      expect(result).toBe(false);
-    });
-    it("should return false if there are more than one correct answers", () => {
-      const result = isMultiAnswerMCQ(multiAnswer);
-      expect(result).toBe(true);
-    });
-  });
-  describe("isShortAnswer", () => {
-    it("should return true if there is a short answer", () => {
-      const result = isShortAnswer(shortAnswer);
-      expect(result).toBe(true);
-    });
-    it("should return false if there is no short answer", () => {
-      const result = isShortAnswer(multiAnswer);
-      expect(result).toBe(false);
-    });
-  });
   describe("pickAnswerComponent", () => {
     it("should return QuizMCQMultiAnswer if isMultiAnswerMCQ is true", () => {
       const result = pickAnswerComponent(multiAnswer);
